use YouTubeEmbed on help page too
